--- src/main/java/com/future/function/qa/model/request/auth/AuthWebRequest.java
@@ -1,17 +1,0 @@
-package com.future.function.qa.model.request.auth;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class AuthWebRequest {
-
-  private String email;
-
-  private String password;
-}
