+++ src/main/java/com/future/function/qa/model/request/core/auth/AuthWebRequest.java
@@ -1,0 +1,17 @@
+package com.future.function.qa.model.request.core.auth;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class AuthWebRequest {
+
+  private String email;
+
+  private String password;
+}
