--- src/main/java/com/future/function/qa/model/request/sticky_note/StickyNoteWebRequest.java
@@ -1,17 +1,0 @@
-package com.future.function.qa.model.request.sticky_note;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class StickyNoteWebRequest {
-
-  private String description;
-
-  private String title;
-}
