--- src/main/java/com/future/function/qa/model/response/auth/AuthWebResponse.java
@@ -1,17 +1,0 @@
-package com.future.function.qa.model.response.auth;
-
-import lombok.Data;
-
-@Data
-public class AuthWebResponse {
-
-  private String avatar;
-
-  private String email;
-
-  private String id;
-
-  private String name;
-
-  private String role;
-}
