+++ src/main/java/com/future/function/qa/model/response/core/auth/AuthWebResponse.java
@@ -1,0 +1,17 @@
+package com.future.function.qa.model.response.core.auth;
+
+import lombok.Data;
+
+@Data
+public class AuthWebResponse {
+
+  private String avatar;
+
+  private String email;
+
+  private String id;
+
+  private String name;
+
+  private String role;
+}
