+++ src/main/java/com/future/function/qa/model/response/core/batch/BatchWebResponse.java
@@ -1,0 +1,19 @@
+package com.future.function.qa.model.response.core.batch;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class BatchWebResponse {
+
+  private String code;
+
+  private String id;
+
+  private String name;
+}
