--- src/main/java/com/future/function/qa/model/response/sticky_note/StickyNoteWebResponse.java
@@ -1,21 +1,0 @@
-package com.future.function.qa.model.response.sticky_note;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class StickyNoteWebResponse {
-
-  private String description;
-
-  private String id;
-
-  private String title;
-
-  private Long updatedAt;
-}
