+++ src/test/resources/features/Core/StickyNote/StickyNote.feature
